+++ README.md
@@ -5,6 +5,8 @@ Juego de tanques para dos personas hecho con Pilas Engine.
 El juego está orientado a enseñar a la gente lo que se puede hacer con Pilas
 en pocos pasos.
 
+Una actualización realizada por Diego Saavedra con la colaboración de (agregar colaborador)
+
 http://pilas-engine.com.ar/
 
 ![](https://raw.github.com/quiqueporta/tank/master/picture.png)
