+++ paso2.py
@@ -1,100 +1,75 @@
 # -*- encoding: utf-8 -*-
 #
-# Creamos los tanques con sus controles asociados.
+# Creamos las escenas y el menú del juego
 
-import random
+import pilasengine
 
-import pilas
-from pilas.escena import Normal
-from pilas.actores import Actor
+pilas = pilasengine.iniciar()
 
+teclas2 = {
+            pilas.simbolos.a: 'izquierda',
+            pilas.simbolos.d: 'derecha',
+            pilas.simbolos.w: 'arriba',
+            pilas.simbolos.s: 'abajo',
+            pilas.simbolos.ALTGR: 'boton',
+        }
 
-class Tanque(Actor):
+class Tanque(pilasengine.actores.Actor):
 
-    def __init__(self, control, imagen):
+    def iniciar(self):
+        self.aprender("Disparar", frecuencia_de_disparo=2)
+        self.aprender("LimitadoABordesDePantalla")
+        self.aprender("PuedeExplotar")
 
-        # Obtenemos la imagen del tanque.
-        imagen_tanque = pilas.imagenes.cargar_imagen(imagen)
-
-        x = random.randrange(-320, 320)
-        y = random.randrange(-240, 240)
-
-        # Iniciamos el actor con la imagen del tanque.
-        super(Tanque, self).__init__(imagen_tanque, x=x, y=y)
-
-        # Establecemos la habilidad de disparar al tanque.
-        self.aprender(pilas.habilidades.Disparar,
-                      control=control,
-                      frecuencia_de_disparo=2)
-
-        # Establecemos la habilidad de moverse.
-        self.aprender(pilas.habilidades.MoverseComoCoche,
-                      control=control,
-                      velocidad_maxima=2,
-                      deceleracion=0.05,
-                      velocidad_rotacion=0.5)
-
-        # Habilidad para que nunca desaparezca de la pantalla.
-        self.aprender(pilas.habilidades.SeMantieneEnPantalla)
-
-        self.aprender(pilas.habilidades.PuedeExplotar)
+    def actualizar(self):
+        pass
 
 
-class Escena_Juego(Normal):
+class Escena_Juego(pilasengine.escenas.Escena):
     """ Escena principal del juego. """
 
     def iniciar(self):
         # Cargamos el fondo del juego.
-        pilas.fondos.Pasto()
+        self.imagen = pilas.fondos.Pasto()
+        tanque1= pilas.actores.Tanque()
+        tanque1.x = 200
+        tanque1.imagen = "images/tanque.png"
+        tanque1.rotacion = 90
+        mi_control = pilas.control.Control(teclas2)
+        tanque1.aprender("MoverseComoCoche", control=mi_control,velocidad_maxima=2, deceleracion=0.05, velocidad_rotacion=0.5)
 
-        self.tanque_J1 = self.crear_tanque(pilas.simbolos.ARRIBA,
-                                           pilas.simbolos.ABAJO,
-                                           pilas.simbolos.IZQUIERDA,
-                                           pilas.simbolos.DERECHA,
-                                           pilas.simbolos.ALTGR,
-                                           "images/tanque.png")
-
-        self.tanque_J2 = self.crear_tanque(pilas.simbolos.w,
-                                           pilas.simbolos.s,
-                                           pilas.simbolos.a,
-                                           pilas.simbolos.d,
-                                           pilas.simbolos.SHIFT,
-                                           "images/tanque2.png")
-
-    def crear_tanque(self, arriba, abajo, izquierda, derecha, disparo, imagen):
-
-        teclas = {izquierda: 'izquierda',
-                  derecha: 'derecha',
-                  arriba: 'arriba',
-                  abajo: 'abajo',
-                  disparo: 'boton'}
-        control = pilas.control.Control(pilas.escena_actual(), teclas)
-        tanque = Tanque(control, imagen)
-        return tanque
+        tanque2 = pilas.actores.Tanque()
+        tanque2.x = -200
+        tanque2.imagen = "images/tanque2.png"
+        tanque2.aprender("MoverseComoCoche", velocidad_maxima=2, deceleracion=0.05, velocidad_rotacion=0.5)
 
 
-class Escena_Menu(Normal):
+class Escena_Menu(pilasengine.escenas.Escena):
     """ Escena del menú del juego. """
-
-    def iniciar_juego(self):
-        pilas.cambiar_escena(Escena_Juego())
-
-    def salir_del_juego(self):
-        pilas.terminar()
 
     def iniciar(self):
         # Cargamos el fondo del juego.
         pilas.fondos.Tarde()
-
-        opciones = [
-            ('Iniciar Juego', self.iniciar_juego),
-            ('Salir', self.salir_del_juego),
+        menu = pilas.actores.Menu(
+        [
+        (u'Iniciar Juego', self.iniciar_juego),
+        (u'Salir', self.salir_del_juego),
         ]
+        )
 
-        pilas.actores.Menu(opciones)
+    def salir_del_juego(self):
+        pilas.terminar()
 
-pilas.iniciar()
+    def iniciar_juego(self):
+        pilas.escenas.Escena_Juego()
 
-pilas.cambiar_escena(Escena_Menu())
+
+pilas.escenas.vincular(Escena_Menu)
+pilas.escenas.vincular(Escena_Juego)
+
+pilas.actores.vincular(Tanque)
+
+pilas.escenas.Escena_Menu()
+
 
 pilas.ejecutar()
